three-address interpreter, restructure three-addr ir, use slingshot
add the "conditional call" pass, quite early, which rewrites `if` into `callif` of a
closure; adjust subsequent passes to use `callif` instead of `if`.

add "threeval" (three + eval), an interpreter for the threeaddr ir. it's properly
tail-recursive, which i think is pretty neat.

use slingshot's `try+` and `throw+` for error handling instead of clojure's `try` and
`throw`, because afaict it's either impossible or at least hugely unergonomic to define
custom exception types in clojure and discriminate on them in `catch` forms.